requirements.txt
asgiref==3.3.1
blurhash-python==1.1.2
cffi==1.14.6
dj-database-url==0.5.0
Django==3.1.3
django-cleanup==5.1.0
django-cors-headers==3.8.0
django-crispy-forms==1.10.0
django-dbbackup==3.3.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2020.4
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
Whoosh==2.7.4
PyYAML~=5.4.1